% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMH.R
\name{get_ll}
\alias{get_ll}
\title{Function to calculate likelihood or 'noisy'/auxillary likelihood.}
\usage{
get_ll(x, lambda, sigma_aux = NULL)
}
\arguments{
\item{x}{data}

\item{lambda}{initial proposal for parameter value}

\item{sigma_aux}{the SD of the auxillary distribution (relative magnitude of noise introduced to the target dist).  Set to NULL to run regular M-H.}
}
\description{
Function to calculate likelihood or 'noisy'/auxillary likelihood.
}
