% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMH.R
\name{runMH}
\alias{runMH}
\title{Run M-H chain}
\usage{
runMH(x, lambda, sigma, N, sigma_aux = NULL)
}
\arguments{
\item{x}{data}

\item{lambda}{initial proposal for parameter value}

\item{sigma}{the SD of the proposal distribution. is not the expected jump size (depends on the posterior, but also the prob that a proposal is accepted, so a function of the posterior)}

\item{N}{the number of iterations of M-H that you'd like to run}

\item{sigma_aux}{the SD of the auxillary distribution (relative magnitude of noise introduced to the target dist).  Set to NULL to run regular M-H.}
}
\description{
Run M-H chain
}
