% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMH.R
\name{stepMH}
\alias{stepMH}
\title{Perform single M-H step}
\usage{
stepMH(x, lambda, sigma, sigma_aux = NULL)
}
\arguments{
\item{x}{data}

\item{lambda}{initial proposal for parameter value}

\item{sigma}{the SD of the proposal distribution. is not the expected jump size (depends on the posterior, but also the prob that a proposal is accepted, so a function of the posterior)}

\item{sigma_aux}{the SD of the auxillary distribution (relative magnitude of noise introduced to the target dist).  Set to NULL to run regular M-H.}
}
\description{
Perform single M-H step
}
